+++ CHANGES.md
@@ -26,6 +26,7 @@ Note that default values of `num_subsets` and `subset_num` are 0 and 1 respectiv
 * Registration of 2d images with aladin and f3d. 
 * When registering, internally the forward displacement is no loner stored, replaced by the forward deformation. The inverse is no longer stored, and is calculated as needed.
 * When adding a shape to a `sirf.STIR.ImageData`, optionally give the number of times to sample a voxel. This is useful when the shape partially - but not completely - fills a voxel.
+* Renamed `Resample` to `Resampler` and `NiftyResample` to `NiftyResampler`
 
 ## v2.2.0
 * Changed CCP PETMR to SyneRBI